+++ chaincode/abstore/go/abstore.go
@@ -1,174 +1,119 @@
-/*
-Copyright IBM Corp. 2016 All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-		 http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
 package main
 
 import (
-	"errors"
+	"encoding/json"
 	"fmt"
-	"strconv"
 	"github.com/hyperledger/fabric-contract-api-go/contractapi"
 )
 
-// ABstore Chaincode implementation
+// User 구조체는 사용자 정보를 나타냅니다.
+type User struct {
+	ID   string `json:"id"`
+	Name string `json:"name"`
+}
+
+// Property 구조체는 부동산 정보를 나타냅니다.
+type Property struct {
+	ID          string `json:"id"`
+	OwnerID     string `json:"ownerId"`
+	Description string `json:"description"`
+}
+
+// ABstore 체인코드의 구조체 정의를 확장합니다.
 type ABstore struct {
 	contractapi.Contract
 }
 
-func (t *ABstore) Init(ctx contractapi.TransactionContextInterface, A string, Aval int, B string, Bval int, C string, Cval int) error {
-	fmt.Println("ABstore Init")
-	var err error
-	// Initialize the chaincode
-	fmt.Printf("Aval = %d, Bval = %d\n, Cval = %d\n", Aval, Bval,Cval)
-	// Write the state to the ledger
-	err = ctx.GetStub().PutState(A, []byte(strconv.Itoa(Aval)))
+// RegisterUser는 새로운 사용자를 등록합니다.
+func (t *ABstore) RegisterUser(ctx contractapi.TransactionContextInterface, userID, name string) error {
+	user := User{
+		ID:   userID,
+		Name: name,
+	}
+	userBytes, err := json.Marshal(user)
 	if err != nil {
 		return err
 	}
 
-	err = ctx.GetStub().PutState(B, []byte(strconv.Itoa(Bval)))
+	return ctx.GetStub().PutState(userID, userBytes)
+}
+
+// RegisterProperty는 새로운 부동산을 등록합니다.
+func (t *ABstore) RegisterProperty(ctx contractapi.TransactionContextInterface, propertyID, ownerID, description string) error {
+	property := Property{
+		ID:          propertyID,
+		OwnerID:     ownerID,
+		Description: description,
+	}
+	propertyBytes, err := json.Marshal(property)
 	if err != nil {
 		return err
 	}
 
-	err = ctx.GetStub().PutState(C, []byte(strconv.Itoa(Cval)))
+	return ctx.GetStub().PutState(propertyID, propertyBytes)
+}
+
+// UpdatePropertyDescription은 부동산 설명을 업데이트합니다.
+func (t *ABstore) UpdatePropertyDescription(ctx contractapi.TransactionContextInterface, propertyID, newDescription string) error {
+	propertyBytes, err := ctx.GetStub().GetState(propertyID)
 	if err != nil {
 		return err
 	}
 
+	if propertyBytes == nil {
+		return fmt.Errorf("property %s does not exist", propertyID)
+	}
 
-	return nil
-}
-
-// Transaction makes payment of X units from A to B
-func (t *ABstore) Invoke(ctx contractapi.TransactionContextInterface, A, B, C string, X int) error {
-	var err error
-	var Aval int
-	var Bval int
-	var Cval int
-	// Get the state from the ledger
-	// TODO: will be nice to have a GetAllState call to ledger
-	Avalbytes, err := ctx.GetStub().GetState(A)
-	if err != nil {
-		return fmt.Errorf("Failed to get state")
-	}
-	if Avalbytes == nil {
-		return fmt.Errorf("Entity not found")
-	}
-	Aval, _ = strconv.Atoi(string(Avalbytes))
-
-	Bvalbytes, err := ctx.GetStub().GetState(B)
-	if err != nil {
-		return fmt.Errorf("Failed to get state")
-	}
-	if Bvalbytes == nil {
-		return fmt.Errorf("Entity not found")
-	}
-	Bval, _ = strconv.Atoi(string(Bvalbytes))
-
-	Cvalbytes, err := ctx.GetStub().GetState(C)
-	if err != nil {
-		return fmt.Errorf("Failed to get state")
-	}
-	if Cvalbytes == nil {
-		return fmt.Errorf("Entity not found")
-	}
-	Cval, _ = strconv.Atoi(string(Cvalbytes))
-	
-	// Perform the execution
-	Aval = Aval - X
-	Bval = Bval + X - ( X / 10 )
-	Cval = Cval + ( X / 10 )
-	fmt.Printf("Aval = %d, Bval = %d, Cval = %d\n", Aval, Bval, Cval)
-
-	// Write the state back to the ledger
-	err = ctx.GetStub().PutState(A, []byte(strconv.Itoa(Aval)))
+	property := Property{}
+	err = json.Unmarshal(propertyBytes, &property)
 	if err != nil {
 		return err
 	}
 
-	err = ctx.GetStub().PutState(B, []byte(strconv.Itoa(Bval)))
+	property.Description = newDescription
+	updatedPropertyBytes, err := json.Marshal(property)
 	if err != nil {
 		return err
 	}
 
-	err = ctx.GetStub().PutState(C, []byte(strconv.Itoa(Cval)))
+	return ctx.GetStub().PutState(propertyID, updatedPropertyBytes)
+}
+
+// QueryProperty는 부동산 정보를 조회합니다.
+func (t *ABstore) QueryProperty(ctx contractapi.TransactionContextInterface, propertyID string) (*Property, error) {
+	propertyBytes, err := ctx.GetStub().GetState(propertyID)
+	if err != nil {
+		return nil, err
+	}
+
+	if propertyBytes == nil {
+		return nil, fmt.Errorf("property %s does not exist", propertyID)
+	}
+
+	property := Property{}
+	err = json.Unmarshal(propertyBytes, &property)
+	if err != nil {
+		return nil, err
+	}
+
+	return &property, nil
+}
+
+// TransferPropertyOwnership은 부동산 소유권을 이전합니다.
+func (t *ABstore) TransferPropertyOwnership(ctx contractapi.TransactionContextInterface, propertyID, newOwnerID string) error {
+	propertyBytes, err := ctx.GetStub().GetState(propertyID)
 	if err != nil {
 		return err
 	}
 
-	return nil
-}
+	if propertyBytes == nil {
+		return fmt.Errorf("property %s does not exist", propertyID)
+	}
 
-// Delete  an entity from state
-func (t *ABstore) Delete(ctx contractapi.TransactionContextInterface, A string) error {
-
-	// Delete the key from the state in ledger
-	err := ctx.GetStub().DelState(A)
+	property := Property{}
+	err = json.Unmarshal(propertyBytes, &property)
 	if err != nil {
-		return fmt.Errorf("Failed to delete state")
+		return err
 	}
 
-	return nil
-}
-
-// Query callback representing the query of a chaincode
-func (t *ABstore) Query(ctx contractapi.TransactionContextInterface, A string) (string, error) {
-	var err error
-	// Get the state from the ledger
-	Avalbytes, err := ctx.GetStub().GetState(A)
-	if err != nil {
-		jsonResp := "{\"Error\":\"Failed to get state for " + A + "\"}"
-		return "", errors.New(jsonResp)
-	}
-
-	if Avalbytes == nil {
-		jsonResp := "{\"Error\":\"Nil amount for " + A + "\"}"
-		return "", errors.New(jsonResp)
-	}
-
-	jsonResp := "{\"Name\":\"" + A + "\",\"Amount\":\"" + string(Avalbytes) + "\"}"
-	fmt.Printf("Query Response:%s\n", jsonResp)
-	return string(Avalbytes), nil
-}
-
-func (t *ABstore) GetAllQuery(ctx contractapi.TransactionContextInterface) ([]string, error) {
-    resultsIterator, err := ctx.GetStub().GetStateByRange("", "")
-    if err != nil {
-        return nil, err
-    }
-    defer resultsIterator.Close()
-    var wallet []string
-    for resultsIterator.HasNext() {
-        queryResponse, err := resultsIterator.Next()
-        if err != nil {
-            return nil, err
-        }
-        jsonResp := "{\"Name\":\"" + string(queryResponse.Key) + "\",\"Amount\":\"" + string(queryResponse.Value) + "\"}"
-        wallet = append(wallet, jsonResp)
-    }
-    return wallet, nil
-}
-
-func main() {
-	cc, err := contractapi.NewChaincode(new(ABstore))
-	if err != nil {
-		panic(err.Error())
-	}
-	if err := cc.Start(); err != nil {
-		fmt.Printf("Error starting ABstore chaincode: %s", err)
-	}
-}
+	property.OwnerID = newOwnerID
